requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bitsandbytes==0.49.2
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
datasets==4.8.4
dill==0.4.1
distro==1.9.0
fastapi==0.135.3
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.72.0
googleapis-common-protos==1.74.0
gradio==6.12.0
gradio_client==2.4.1
groovy==0.1.2
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-gradio==0.3.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.10.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
numpy==2.4.4
orjson==3.11.8
packaging==26.0
pandas==3.0.2
peft==0.18.1
pillow==12.2.0
platformdirs==4.9.6
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==14.3.4
safehttpx==0.1.7
safetensors==0.7.0
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.57.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.11.0
tqdm==4.67.3
transformers==5.5.3
trl==1.0.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
wandb==0.25.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
